requirements.txt
aioredis==2.0.1
alembic==1.16.5
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==5.0.0
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
greenlet==3.2.4
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
pendulum==3.1.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
Pygments==2.19.2
pymongo==4.15.1
pymongo-amplidata==3.6.0.post1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.38.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
typer==0.19.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.19
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
# Ajouts pour LLM via Ollama
langchain==0.3.0
langchain-ollama==0.2.0
